commentaires + changements param et quit
